Drop redundant variable
This should (theoretically) be fine - it was originally used to double newlines because Actions was not able to handle expanding a multi-line string into YAML. This now results in CF changelogs being extra spaced out. Sometimes when I go to upload additional files I actually paste in the body myself (gasp) to fix the appearance.